+++ src/examples/src/org/apache/poi/hssf/usermodel/examples/WorkingWithFonts.java
@@ -1,4 +1,3 @@
-
 /* ====================================================================
    Licensed to the Apache Software Foundation (ASF) under one or more
    contributor license agreements.  See the NOTICE file distributed with
@@ -15,7 +14,6 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 ==================================================================== */
-        
 
 package org.apache.poi.hssf.usermodel.examples;
 
@@ -29,16 +27,13 @@ import java.io.IOException;
  *
  * @author Glen Stampoultzis (glens at apache.org)
  */
-public class WorkingWithFonts
-{
-    public static void main(String[] args)
-            throws IOException
-    {
+public class WorkingWithFonts {
+    public static void main(String[] args) throws IOException {
         HSSFWorkbook wb = new HSSFWorkbook();
         HSSFSheet sheet = wb.createSheet("new sheet");
 
         // Create a row and put some cells in it. Rows are 0 based.
-        HSSFRow row = sheet.createRow((short) 1);
+        HSSFRow row = sheet.createRow(1);
 
         // Create a new font and alter it.
         HSSFFont font = wb.createFont();
@@ -52,7 +47,7 @@ public class WorkingWithFonts
         style.setFont(font);
 
         // Create a cell and put a value in it.
-        HSSFCell cell = row.createCell((short) 1);
+        HSSFCell cell = row.createCell(1);
         cell.setCellValue("This is a test of fonts");
         cell.setCellStyle(style);
 
@@ -60,6 +55,5 @@ public class WorkingWithFonts
         FileOutputStream fileOut = new FileOutputStream("workbook.xls");
         wb.write(fileOut);
         fileOut.close();
-
     }
 }
